+++ src/components/Navbar/components/DesktopNav.module.css
@@ -1,0 +1,32 @@
+.desktopNav {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 1rem;
+
+  & .navLinks {
+    max-width: 60vw;
+    width: 100%;
+  }
+
+  & ul {
+    display: flex;
+
+    list-style-type: none;
+    padding-inline: 0;
+
+    & li {
+      margin-inline-end: 3rem;
+    }
+
+    & a {
+      color: #2f2f2f;
+    }
+  }
+}
+
+.darkNav {
+  & ul > li > a {
+    color: #fffff2;
+  }
+}
