+++ src/components/Navbar/index.module.css
@@ -1,53 +1,11 @@
-.component {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-
-  padding: 0.5rem 1rem;
-
-  & .button {
-    width: 1.25rem;
-    height: 1.25rem;
-  }
-
-  & .navOpen {
-    background-color: #ffffff;
-    display: block;
-    position: fixed;
-    height: 100%;
-    width: 100%;
-
-    @media screen and (min-width: 768px) {
-      /* bottom: auto;
-      height: auto; */
-    }
-  }
-
-  & a {
-    color: #2f2f2f;
-    margin-inline-end: 1rem;
-    text-decoration: none;
+.mobile {
+  @media screen and (min-width: 768px) {
+    display: none;
   }
 }
 
-.nav {
-  display: none;
-
-  & ul {
-    list-style-type: none;
-    padding-inline: 0;
-  }
-
-  & ul > li {
-    font-size: 1.25rem;
-    margin-block: 1rem;
-  }
-}
-
-.darkNav {
-  background-color: #1d2530 !important;
-
-  & a {
-    color: #fffff2 !important;
+.desktop {
+  @media (max-width: 767px) {
+    display: none;
   }
 }
